Improved

+++ fotoshop/src/main/java/net/drs/fotoshop/api/userdetails/IUserDetails.java
@@ -21,19 +21,20 @@ public interface IUserDetails  {
 	List<User> getAllUsers();
 	
 	// get the information of specific user - based on user id
-	User getUserById(int userid);
+	User getUserById(Long userid);
+	
 	
 	// get all Active users
 	List<User> getAllActiveUsers();
 	
 	//check if the user is active or not, based on userid
-	boolean isUserActive(int userId);
+	boolean isUserActive(Long userId);
 	
 	// Activate User
-	boolean activeteUser(int userId);
+	boolean activeteUser(Long userId);
 	
 	// Deactivate User / Delete User
-	boolean deactiveUser(int userId);
+	boolean deactiveUser(Long userId);
 	
 	
 	//if the user need to update the details..
